requirements.txt
fastapi==0.88.0
mysql_connector_repackaged==0.3.1
python-dotenv==0.21.0
